refactor(client): improve size
