+++ packages/client/src/components/CommentCard.vue
@@ -2,6 +2,7 @@
   <div :id="comment.objectId" class="wl-item">
     <div class="wl-user" aria-hidden="true">
       <img v-if="comment.avatar" :src="comment.avatar" />
+
       <VerifiedIcon v-if="comment.type" />
     </div>
 
@@ -12,9 +13,10 @@
           class="wl-nick"
           :href="link"
           target="_blank"
-          rel="nofollow noreferrer"
+          rel="noopener noreferrer"
           >{{ comment.nick }}</a
         >
+
         <span v-else class="wl-nick">{{ comment.nick }}</span>
 
         <span
@@ -22,26 +24,32 @@
           class="wl-badge"
           v-text="locale.admin"
         />
+
         <span v-if="comment.label" class="wl-badge" v-text="comment.label" />
+
         <span v-if="comment.sticky" class="wl-badge" v-text="locale.sticky" />
+
         <span
           v-if="comment.level !== undefined && comment.level >= 0"
           :class="`wl-badge level${comment.level}`"
           v-text="locale[`level${comment.level}`] || `Level ${comment.level}`"
         />
+
         <span class="wl-time" v-text="time" />
 
         <div class="wl-comment-actions">
           <button
             v-if="isAdmin || isOwner"
+            type="button"
             class="wl-edit"
-            @click="$emit('edit', comment)"
+            @click="() => $emit('edit', comment)"
           >
             <EditIcon />
           </button>
 
           <button
             v-if="isAdmin || isOwner"
+            type="button"
             class="wl-delete"
             @click="$emit('delete', comment)"
           >
@@ -49,15 +57,18 @@
           </button>
 
           <button
+            type="button"
             class="wl-like"
             :title="like ? locale.cancelLike : locale.like"
             @click="$emit('like', comment)"
           >
             <LikeIcon :active="like" />
+
             <span v-if="'like' in comment" v-text="comment.like" />
           </button>
 
           <button
+            type="button"
             class="wl-reply"
             :class="{ active: isReplyingCurrent }"
             :title="isReplyingCurrent ? locale.cancelReply : locale.reply"
@@ -67,6 +78,7 @@
           </button>
         </div>
       </div>
+
       <div class="wl-meta" aria-hidden="true">
         <span
           v-if="comment.addr"
@@ -74,12 +86,14 @@
           :data-value="comment.addr"
           v-text="comment.addr"
         />
+
         <span
           v-if="comment.browser"
           class="wl-browser"
           :data-value="comment.browser"
           v-text="comment.browser"
         />
+
         <span
           v-if="comment.os"
           class="wl-os"
@@ -100,6 +114,7 @@
           <button
             v-for="status in commentStatus"
             :key="status"
+            type="submit"
             :class="`wl-btn wl-${status}`"
             :disabled="comment.status === status"
             @click="$emit('status', { status, comment })"
@@ -109,6 +124,7 @@
 
         <button
           v-if="isAdmin && !comment.rid"
+          type="submit"
           class="wl-btn wl-sticky"
           @click="$emit('sticky', comment)"
         >
@@ -128,12 +144,15 @@
           :reply-id="reply?.objectId"
           :reply-user="comment.nick"
           :root-id="rootId"
-          @submit="$emit('submit', $event)"
           @cancel-reply="$emit('reply', null)"
           @cancel-edit="$emit('edit', null)"
+          @submit="$emit('submit', $event)"
         />
       </div>
+
       <div v-if="comment.children" class="wl-quote">
+        <!-- FIXME: This is a upstream bug -->
+        <!-- eslint-disable-next-line vue/no-undef-components -->
         <CommentCard
           v-for="child in comment.children"
           :key="child.objectId"
@@ -141,21 +160,21 @@
           :reply="reply"
           :edit="edit"
           :root-id="rootId"
-          @reply="$emit('reply', $event)"
-          @submit="$emit('submit', $event)"
-          @like="$emit('like', $event)"
-          @edit="$emit('edit', $event)"
           @delete="$emit('delete', $event)"
+          @edit="$emit('edit', $event)"
+          @like="$emit('like', $event)"
+          @reply="$emit('reply', $event)"
           @status="$emit('status', $event)"
           @sticky="$emit('sticky', $event)"
+          @submit="$emit('submit', $event)"
         />
       </div>
     </div>
   </div>
 </template>
 
-<script lang="ts">
-import { computed, defineComponent, inject } from 'vue';
+<script setup lang="ts">
+import { computed, inject } from 'vue';
 import CommentBox from './CommentBox.vue';
 import {
   DeleteIcon,
@@ -163,97 +182,86 @@ import {
   ReplyIcon,
   EditIcon,
   VerifiedIcon,
-} from './Icons';
-import { isLinkHttp } from '../utils';
-import { useTimeAgo, useLikeStorage, useUserInfo } from '../composables';
+} from './Icons.js';
+import { getTimeAgo, isLinkHttp } from '../utils';
+import { useLikeStorage, useUserInfo } from '../composables/index.js';
 
-import type { ComputedRef, PropType } from 'vue';
-import type { WalineConfig } from '../utils';
-import type { WalineComment, WalineCommentStatus } from '../typings';
+import type { ComputedRef } from 'vue';
+import type { WalineConfig } from '../utils/index.js';
+import type { WalineComment, WalineCommentStatus } from '../typings/index.js';
+import { useNow } from '@vueuse/core';
+
+const props = withDefaults(
+  defineProps<{
+    /**
+     * Comment data
+     */
+    comment: WalineComment;
+    /**
+     * Current comment to be edited
+     */
+    edit?: WalineComment | null;
+    /**
+     * Root comment id
+     */
+    rootId: string;
+    /**
+     * Current comment to be replied
+     */
+    reply?: WalineComment | null;
+  }>(),
+  {
+    edit: null,
+    reply: null,
+  }
+);
+
+defineEmits<{
+  (event: 'submit', comment: WalineComment): void;
+  (event: 'delete', comment: WalineComment): void;
+  (event: 'edit', comment: WalineComment | null): void;
+  (event: 'like', comment: WalineComment): void;
+  (
+    event: 'status',
+    statusInfo: { status: WalineCommentStatus; comment: WalineComment }
+  ): void;
+  (event: 'sticky', comment: WalineComment): void;
+  (event: 'reply', comment: WalineComment | null): void;
+}>();
 
 const commentStatus: WalineCommentStatus[] = ['approved', 'waiting', 'spam'];
 
-export default defineComponent({
-  components: {
-    CommentBox,
-    DeleteIcon,
-    LikeIcon,
-    ReplyIcon,
-    EditIcon,
-    VerifiedIcon,
-  },
+const config = inject<ComputedRef<WalineConfig>>('config')!;
+const likes = useLikeStorage();
+const now = useNow();
+const userInfo = useUserInfo();
 
-  props: {
-    comment: {
-      type: Object as PropType<WalineComment>,
-      required: true,
-    },
-    rootId: {
-      type: String,
-      required: true,
-    },
-    reply: {
-      type: Object as PropType<WalineComment | null>,
-      default: null,
-    },
-    edit: {
-      type: Object as PropType<WalineComment | null>,
-      default: null,
-    },
-  },
+const locale = computed(() => config.value.locale);
 
-  emits: ['submit', 'reply', 'like', 'delete', 'status', 'sticky', 'edit'],
+const link = computed(() => {
+  const { link } = props.comment;
 
-  setup(props) {
-    const config = inject<ComputedRef<WalineConfig>>(
-      'config'
-    ) as ComputedRef<WalineConfig>;
-    const likes = useLikeStorage();
-    const userInfo = useUserInfo();
-
-    const locale = computed(() => config.value.locale);
-
-    const link = computed(() => {
-      const { link } = props.comment;
-
-      return link ? (isLinkHttp(link) ? link : `https://${link}`) : '';
-    });
-
-    const like = computed(() => likes.value.includes(props.comment.objectId));
-
-    const time = useTimeAgo(props.comment.insertedAt, locale.value);
-
-    const isAdmin = computed(() => userInfo.value.type === 'administrator');
-
-    const isOwner = computed(
-      () =>
-        props.comment.user_id &&
-        userInfo.value.objectId === props.comment.user_id
-    );
-
-    const isReplyingCurrent = computed(
-      () => props.comment.objectId === props.reply?.objectId
-    );
-
-    const isEditingCurrent = computed(
-      () => props.comment.objectId === props.edit?.objectId
-    );
-
-    return {
-      config,
-      locale,
-
-      isReplyingCurrent,
-      isEditingCurrent,
-      link,
-      like,
-      time,
-
-      isAdmin,
-      isOwner,
-
-      commentStatus,
-    };
-  },
+  return link ? (isLinkHttp(link) ? link : `https://${link}`) : '';
 });
+
+const like = computed(() => likes.value.includes(props.comment.objectId));
+
+const time = computed(() =>
+  getTimeAgo(props.comment.insertedAt, now.value, locale.value)
+);
+
+const isAdmin = computed(() => userInfo.value.type === 'administrator');
+
+const isOwner = computed(
+  () =>
+    props.comment.user_id && userInfo.value.objectId === props.comment.user_id
+);
+
+const isReplyingCurrent = computed(
+  () => props.comment.objectId === props.reply?.objectId
+);
+
+const isEditingCurrent = computed(
+  () => props.comment.objectId === props.edit?.objectId
+);
 </script>
