+++ packages/client/src/composables/index.ts
@@ -1,6 +1,5 @@
 export * from './inputs';
 export * from './like';
 export * from './recaptchaV3';
-export * from './timeAgo';
 export * from './vote';
 export * from './userInfo';
