+++ packages/client/src/define.ts
@@ -1,0 +1,3 @@
+declare const SHOULD_VALIDATE: boolean;
+
+export const shouldValidate = SHOULD_VALIDATE;
