+++ packages/client/src/utils/image.ts
@@ -1,7 +1,7 @@
 const isImage = (item: DataTransferItem): boolean =>
   item.type.includes('image');
 
-export const getImagefromDataTransfer = (
+export const getImageFromDataTransfer = (
   items: DataTransferItemList
 ): File | null => {
   const image = Array.from(items).find(isImage);
