+++ packages/client/template/init.html
@@ -19,6 +19,7 @@
         path: '/',
         comment: true,
         pageview: true,
+        reaction: true,
         serverURL: 'https://walinejs.comment.lithub.cc',
         emoji: [
           '//unpkg.com/@waline/emojis@1.1.0/bilibili',
