+++ packages/client/tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
+    "noImplicitAny": false,
+    "noUnusedParameters": false,
     "jsx": "preserve",
     "module": "ESNext",
     "rootDir": "./src"
